Rename package message to sms
