+++ sms/message.go
@@ -1,4 +1,4 @@
-package message
+package sms
 
 import (
 	"errors"
